[ORO-0] Refactor and error output when it cannot find llc.
