+++ scripts/kernelCompile.py
@@ -2,23 +2,9 @@ import json
 import os
 import subprocess
 import re
+from enumArch import enumArch
 
-def toNumber( arch ):
-    return int(arch[3:], 16)
-
-def enumArch( minArch ):
-    process = subprocess.Popen(['llc', '-march=amdgcn', '-mcpu=help'], shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE )
-    output, errors = process.communicate()
-    lines = output.decode('utf-8').splitlines() + errors.decode('utf-8').splitlines()
-
-    arches = []
-    for line in lines:
-        result = re.match("\s+(gfx[0-9a-f]+).*processor.", line)
-        if result:
-            arch = result.group(1)
-            if toNumber(minArch) <= toNumber(arch):
-                arches.append( arch )
-    return arches
+#for powershell, $ENV:Path+=";..\..\hipsdk\bin"
 
 def getGpuList():
   f = open("amdGpuList.json")
@@ -30,7 +16,7 @@ ps = []
 def compile( index ):
 	if index == 0 :
 		command = [
-			"..\\..\\hipsdk\\bin\\hipcc",
+			"hipcc",
 			"-x", "hip", "..\ParallelPrimitives\RadixSortKernels.h", "-O3", "-std=c++17", "-ffast-math", "--cuda-device-only", "--genco", "-I../", "-include", "hip/hip_runtime.h", "-parallel-jobs=15"]
 		#command.append( "--offload-arch=gfx1100" )
 		for i in enumArch( "gfx900" ):
